register controller bugs fixed

+++ src/utils/cloudinary.js
@@ -1,34 +1,36 @@
-import {v2 as cloudinary} from "cloudinary";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-
-cloudinary.config({ 
-         cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
     api_secret: process.env.CLOUDINARY_API_SECRET
-    });
+});
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null;
 
         const result = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto",
+            resource_type: "auto"
         });
 
         console.log(
             "File uploaded on Cloudinary successfully",
-            result
+            result.url
         );
+
+        fs.unlinkSync(localFilePath);
 
         return result;
 
     } catch (error) {
-      fs.unlinkSync(localFilePath); // Delete the local file if upload fails
-        console.error(
-            "Error uploading file to Cloudinary:",
-            error
-        );
+
+        if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
+        console.error("Error uploading file to Cloudinary:", error);
 
         return null;
     }
